requirements.txt
uwsgi>=2.0
Flask>=0.10
mysql-connector-python-rf
